// src/app/model/quotation.ts
export interface PostQuotation {
    ref_number: Number;
    valid_from: String;
    valid_to: String;
    line: Number;
    pol_office: Number;
    pod_office: Number;
    party: Number;
    currency: Number;
    licd: Number;
    licd_pol_trans_mode: Number;
    pol: Number;
    pod: Number;
    dicd_pod_trans_mode: Number;
    dicd: Number;
    cargo_type: Number;
    cargo_sub_type: Number;
    un_no: Number;
    stuffing: Number;
    de_stuffing: Number;
  }
  export interface lineInfo {
    id: Number;
  }